requirements.txt
click>=8.0
requests>=2.28
python-dotenv>=1.0
loguru>=0.7
tabulate>=0.8
